conflate common auth functions
