+++ swift_browser_ui/request/api.py
@@ -7,7 +7,7 @@ import aiohttp.web
 from asyncpg import InterfaceError
 
 
-from swift_browser_ui.request.db import handle_dropped_connection
+import swift_browser_ui.common.common_db as common_db
 
 
 MODULE_LOGGER = logging.getLogger("api")
@@ -23,7 +23,7 @@ async def handle_share_request_post(request: aiohttp.web.Request) -> aiohttp.web
     try:
         await request.app["db_conn"].add_request(user, container, owner)
     except InterfaceError:
-        handle_dropped_connection(request)
+        common_db.handle_dropped_connection(request)
 
     return aiohttp.web.json_response(
         {"container": container, "user": user, "owner": owner, "date": None}
@@ -39,7 +39,7 @@ async def handle_user_owned_request_listing(
     try:
         ret = await request.app["db_conn"].get_request_owned(user)
     except InterfaceError:
-        handle_dropped_connection(request)
+        common_db.handle_dropped_connection(request)
 
     MODULE_LOGGER.log(logging.DEBUG, f"Returning requests owned by user: {ret}")
 
@@ -55,7 +55,7 @@ async def handle_user_made_request_listing(
     try:
         ret = await request.app["db_conn"].get_request_made(user)
     except InterfaceError:
-        handle_dropped_connection(request)
+        common_db.handle_dropped_connection(request)
 
     MODULE_LOGGER.log(logging.DEBUG, f"Returning requests made by user: {ret}")
 
@@ -71,7 +71,7 @@ async def handle_container_request_listing(
     try:
         ret = await request.app["db_conn"].get_request_container(container)
     except InterfaceError:
-        handle_dropped_connection(request)
+        common_db.handle_dropped_connection(request)
 
     MODULE_LOGGER.log(logging.DEBUG, f"Returning container shared requests: {ret}")
 
@@ -89,7 +89,7 @@ async def handle_user_share_request_delete(
     try:
         await request.app["db_conn"].delete_request(container, owner, user)
     except InterfaceError:
-        handle_dropped_connection(request)
+        common_db.handle_dropped_connection(request)
 
     MODULE_LOGGER.log(logging.DEBUG, f"Deleted {container} for owner {owner}")
 
@@ -114,7 +114,7 @@ async def handle_user_add_token(request: aiohttp.web.Request) -> aiohttp.web.Res
     try:
         await request.app["db_conn"].add_token(project, token, identifier)
     except InterfaceError:
-        handle_dropped_connection(request)
+        common_db.handle_dropped_connection(request)
 
     return aiohttp.web.Response(status=200)
 
@@ -127,7 +127,7 @@ async def handle_user_delete_token(request: aiohttp.web.Request) -> aiohttp.web.
     try:
         await request.app["db_conn"].revoke_token(project, identifier)
     except InterfaceError:
-        handle_dropped_connection(request)
+        common_db.handle_dropped_connection(request)
 
     MODULE_LOGGER.log(logging.DEBUG, f"Deleted {identifier} for project {project}")
 
@@ -141,7 +141,7 @@ async def handle_user_list_tokens(request: aiohttp.web.Request) -> aiohttp.web.R
     try:
         tokens = await request.app["db_conn"].get_tokens(project)
     except InterfaceError:
-        handle_dropped_connection(request)
+        common_db.handle_dropped_connection(request)
 
     # Return only the identifiers
     return aiohttp.web.json_response([rec["identifier"] for rec in tokens])
