+++ swift_browser_ui/sharing/server.py
@@ -12,6 +12,7 @@ import uvloop
 
 import swift_browser_ui.common.common_middleware
 import swift_browser_ui.common.common_handlers
+import swift_browser_ui.common.common_util
 
 from swift_browser_ui.sharing.api import (
     has_access_handler,
@@ -27,10 +28,6 @@ from swift_browser_ui.sharing.api import (
     handle_health_check,
 )
 from swift_browser_ui.sharing.db import DBConn
-from swift_browser_ui.sharing.auth import (
-    read_in_keys,
-    handle_validate_authentication,
-)
 
 
 logging.basicConfig(level=logging.DEBUG)
@@ -56,7 +53,7 @@ async def init_server() -> aiohttp.web.Application:
         middlewares=[
             swift_browser_ui.common.common_middleware.add_cors,  # type:ignore
             swift_browser_ui.common.common_middleware.check_db_conn,  # type:ignore
-            handle_validate_authentication,  # type:ignore
+            swift_browser_ui.common.common_middleware.handle_validate_authentication,  # type:ignore
             swift_browser_ui.common.common_middleware.catch_uniqueness_error,  # type:ignore
         ]
     )
@@ -98,7 +95,7 @@ async def init_server() -> aiohttp.web.Application:
     )
 
     app.on_startup.append(resume_on_start)
-    app.on_startup.append(read_in_keys)
+    app.on_startup.append(swift_browser_ui.common.common_util.read_in_keys)
     app.on_shutdown.append(save_on_shutdown)
 
     return app
