+++ swift_browser_ui/upload/auth.py
@@ -3,14 +3,11 @@
 
 import os
 import hashlib
-import typing
-import hmac
-import time
-import secrets
 
 import aiohttp.web
 
 import swift_browser_ui.common.types
+import swift_browser_ui.common.signature
 from swift_browser_ui.ui._convenience import initiate_os_session
 
 
@@ -35,32 +32,6 @@ async def handle_login(request: aiohttp.web.Request) -> aiohttp.web.StreamRespon
         raise aiohttp.web.HTTPUnauthorized(reason="Login token or project missing")
 
 
-async def read_in_keys(app: aiohttp.web.Application) -> None:
-    """Read in keys to the application."""
-    keys = os.environ.get("SWIFT_UI_API_AUTH_TOKENS", None)
-    app["tokens"] = keys.split(",") if keys is not None else []
-    if app["tokens"]:
-        app["tokens"] = [token.encode("utf-8") for token in app["tokens"]]
-
-
-async def test_signature(
-    tokens: typing.List[bytes],
-    signature: str,
-    message: str,
-    validity: str,
-) -> bool:
-    """Validate signature against the given tokens."""
-    # Check signature expiration
-    if int(validity) < time.time():
-        raise aiohttp.web.HTTPUnauthorized(reason="Signature expired")
-    byte_message = message.encode("utf-8")
-    for token in tokens:
-        digest = hmac.new(token, byte_message, digestmod="sha256").hexdigest()
-        if secrets.compare_digest(digest, signature):
-            return True
-    raise aiohttp.web.HTTPUnauthorized(reason="Missing valid query signature")
-
-
 @aiohttp.web.middleware
 async def handle_validate_authentication(
     request: aiohttp.web.Request,
@@ -76,6 +47,8 @@ async def handle_validate_authentication(
             reason="Query string missing validity or signature"
         )
 
-    await test_signature(request.app["tokens"], signature, validity + path, validity)
+    await swift_browser_ui.common.signature.test_signature(
+        request.app["tokens"], signature, validity + path, validity
+    )
 
     return await handler(request)
