+++ tests/common/test_common_middleware.py
@@ -2,13 +2,14 @@
 
 
 import types
+import unittest.mock
 
 import asynctest
 import aiohttp.web
 
 import asyncpg.exceptions
 
-import swift_browser_ui.common.common_middleware
+import swift_browser_ui.common.common_middleware as middle
 
 
 class CommonMiddlewareTestCase(asynctest.TestCase):
@@ -22,14 +23,13 @@ class CommonMiddlewareTestCase(asynctest.TestCase):
                 "path": "/test",
             }
         )
+        self.app_mock = {}
         self.mock_handler = asynctest.CoroutineMock(return_value=aiohttp.web.Response())
         super().setUp()
 
     async def test_add_cors(self):
         """Test CORS header addition middleware."""
-        resp = await swift_browser_ui.common.common_middleware.add_cors(
-            self.mock_request, self.mock_handler
-        )
+        resp = await middle.add_cors(self.mock_request, self.mock_handler)
         self.assertEqual(
             resp.headers["Access-Control-Allow-Origin"],
             self.mock_request.headers["origin"],
@@ -39,24 +39,18 @@ class CommonMiddlewareTestCase(asynctest.TestCase):
         """Test database connection guard on failure."""
         self.mock_request.app["db_conn"].conn = None
         with self.assertRaises(aiohttp.web.HTTPServiceUnavailable):
-            await swift_browser_ui.common.common_middleware.check_db_conn(
-                self.mock_request, self.mock_handler
-            )
+            await middle.check_db_conn(self.mock_request, self.mock_handler)
 
     async def test_check_db_conn_existing(self):
         """Test database connection guard on success."""
         self.mock_request.app["db_conn"].conn = "placeholder"
-        resp = await swift_browser_ui.common.common_middleware.check_db_conn(
-            self.mock_request, self.mock_handler
-        )
+        resp = await middle.check_db_conn(self.mock_request, self.mock_handler)
         self.assertIsInstance(resp, aiohttp.web.Response)
 
     async def test_check_db_conn_no_database(self):
         """Test database connection guard middleware without database."""
         self.mock_request.app["db_conn"] = None
-        resp = await swift_browser_ui.common.common_middleware.check_db_conn(
-            self.mock_request, self.mock_handler
-        )
+        resp = await middle.check_db_conn(self.mock_request, self.mock_handler)
         self.assertIsInstance(resp, aiohttp.web.Response)
 
     async def test_catch_uniqueness_error(self):
@@ -65,6 +59,47 @@ class CommonMiddlewareTestCase(asynctest.TestCase):
             side_effect=asyncpg.exceptions.UniqueViolationError
         )
         with self.assertRaises(aiohttp.web.HTTPConflict):
-            await swift_browser_ui.common.common_middleware.catch_uniqueness_error(
+            await middle.catch_uniqueness_error(
                 self.mock_request, unique_violating_handler
             )
+
+
+class HandleValidateAuthTest(asynctest.TestCase):
+    """Auth middleware tests."""
+
+    async def test_handle_validate_authentication_success(self):
+        """Test authentication validation handler success."""
+        t_singature_mock = asynctest.CoroutineMock()
+        t_signature_patch = unittest.mock.patch(
+            "swift_browser_ui.common.signature.test_signature", t_singature_mock
+        )
+
+        handler_mock = asynctest.CoroutineMock()
+        request_mock = types.SimpleNamespace(
+            **{
+                "app": {
+                    "tokens": ["awefi"],
+                },
+                "query": {"signature": "a", "valid": "b"},
+                "match_info": {},
+                "url": types.SimpleNamespace(**{"path": "c"}),
+                "path": "/health",
+            }
+        )
+
+        with t_signature_patch:
+            await middle.handle_validate_authentication(request_mock, handler_mock)
+
+    async def test_handle_validate_authentication_failure(self):
+        """Test authentication validation handler failure."""
+        handler_mock = asynctest.CoroutineMock()
+        request_mock = types.SimpleNamespace(
+            **{
+                "query": {"signature": "a", "vali": "b"},
+                "match_info": {},
+                "url": types.SimpleNamespace(**{"path": "c"}),
+                "path": "/health",
+            }
+        )
+        with self.assertRaises(aiohttp.web.HTTPClientError):
+            await middle.handle_validate_authentication(request_mock, handler_mock)
