+++ tests/sharing/test_api.py
@@ -165,7 +165,7 @@ class APILostDatabaseConnectionClass(asynctest.TestCase):
             )
         )
         self.patch_handle_dropped_connection = unittest.mock.patch(
-            "swift_browser_ui.sharing.api.handle_dropped_connection",
+            "swift_browser_ui.sharing.api.common_db.handle_dropped_connection",
             new=self.handle_dropped_connection_mock,
         )
